Extract src/caffeinate.ts — keep-awake
The split button, its presets, the timer, the agents-busy poll and the picker
popover, with its own handlers. 190 lines; main.ts 2,721 -> 2,561. Diffs
identical against HEAD at 163 vs 163 lines.

Three hooks, all belonging to the layer that owns the footer: closeFootMenus,
renderFoot, renderAll. reconcileCaf stays the single choke point and is still
called from renderAll.

Dropped an `isAgent` import the extraction appeared to need: cafAgentsBusy tests
`s.kind === "shell"` directly rather than going through the helper. Left as-is —
it is pre-existing, and rewriting it is a behaviour-adjacent change that belongs
in the dead-code sweep, not in a move.

Smoked through the real caret button: the popover opens with all four presets and
four durations, picking one closes it and persists (cc-caffeinate display ->
system). The cup glyph does not light up, which is the missing Tauri backend
rather than a regression — reconcileCaf invokes the backend before repainting, so
in a browser the assertion never starts.
